// front-end/src/app/models/VillesModel.ts
// Generated by https://quicktype.io

export interface Villes {
    _embedded: Embedded;
    _links:    VillesLinks;
    page:      Page;
}

export interface Embedded {
    villes: Ville[];
}

export interface Ville {
    name:      string;
    longitude: number;
    latitude:  number;
    altitude:  number;
    _links:    VilleLinks;
}

export interface VilleLinks {
    self:               Profile;
    ville:              Profile;
    activitesSportives: Profile;
}

export interface Profile {
    href: string;
}

export interface VillesLinks {
    self:    Profile;
    profile: Profile;
}

export interface Page {
    size:          number;
    totalElements: number;
    totalPages:    number;
    number:        number;
}