// front-end/src/app/models/ActivitesSportivesModel.ts
// Generated by https://quicktype.io

export interface ActivitesSportives {
    _embedded: Embedded;
    _links:    ActivitesSportivesLinks;
    page:      Page;
}

export interface Embedded {
    activiteSportives: ActiviteSportive[];
}

export interface ActiviteSportive {
    name:           string;
    longitude:      number;
    latitude:       number;
    altitude:       number;
    nbrEmplacement: number;
    _links:         ActiviteSportiveLinks;
}

export interface ActiviteSportiveLinks {
    self:             Profile;
    activiteSportive: Profile;
    ville:            Profile;
    emplacements:     Profile;
}

export interface Profile {
    href: string;
}

export interface ActivitesSportivesLinks {
    self:    Profile;
    profile: Profile;
}

export interface Page {
    size:          number;
    totalElements: number;
    totalPages:    number;
    number:        number;
}