// front-end/src/app/models/EmplacementsModel.ts
// Generated by https://quicktype.io

import { SeanceElement } from "./SeanceProjection";

export interface Emplacements {
  _embedded: Embedded;
  _links: EmplacementsLinks;
  page: Page;
  
}

export interface Embedded {
  emplacements: Emplacement[];
}

export interface Emplacement {
  id:string,
  name: string;
  nbrPlace: number;
  photo:string
  _links: EmplacementLinks;
  seancesProjections:SeanceElement[]
}

export interface EmplacementLinks {
  self: Profile;
  emplacement: Profile;
  seances: Profile;
  activiteSportive: Profile;
  places: Profile;
}

export interface Profile {
  href: string;
}

export interface EmplacementsLinks {
  self: Profile;
  profile: Profile;
}

export interface Page {
  size: number;
  totalElements: number;
  totalPages: number;
  number: number;
}