// front-end/src/app/models/ReservationsProjection.ts
// Generated by https://quicktype.io

import { Place } from './PlacesModel';

export interface ReservationsProjection {
  _embedded: Embedded;
  _links: ReservationsProjectionLinks;
  page: Page;
}

export interface Embedded {
  reservations: ReservationElement[];
}

export interface ReservationElement {
  prix: number;
  nomClient: string;
  place: Place;
  codePaiement: number;
  reserve: boolean;
  id: string;
  selected: boolean;
  _links: ReservationLinks;
}

export interface ReservationLinks {
  self: First;
  reservation: SeanceClass;
  place: First;
  seance: SeanceClass;
}

export interface First {
  href: string;
}

export interface SeanceClass {
  href: string;
  templated: boolean;
}

export interface ReservationsProjectionLinks {
  first: First;
  self: First;
  next: First;
  last: First;
  profile: First;
}

export interface Page {
  size: number;
  totalElements: number;
  totalPages: number;
  number: number;
}